requirements.txt
pandas
python-dotenv
requests
pytelegrambotapi
openpyxl
pysondb